refactor(bookmarks): split tag field modules
