+++ docs/superpowers/specs/2026-08-01-bookmark-form-error-ownership-design.md
@@ -117,7 +117,7 @@ fieldの表示メッセージは次の優先順位で決める。
 
 ## 設計判断のコメント
 
-実装では、コードだけを読んだときに別の実装へ変更されやすい判断に、読み手が前提と理由を追える日本語コメントで`Why?`を残す。コメントの長さは判断の複雑さに合わせる。コメントは処理の説明ではなく、責務境界と採用理由を説明する。
+実装では、コードだけを読んだときに別の実装へ変更されやすい判断に、読み手が前提と理由を追える日本語コメントで設計理由を残す。コメントの長さは判断の複雑さに合わせる。コメントは処理の説明ではなく、責務境界と採用理由を説明する。
 
 コメントを書くときは、日本語技術文書の規範を使い、一文ごとの意味、段落の論理、用語の一貫性、読み手が保持する情報量を点検する。短くすること自体を目的にせず、将来の開発者が判断を誤って変更しないために必要な説明を残す。
