--- src/features/bookmarks/components/bookmark-tag-field.tsx
@@ -1,401 +1,0 @@
-import { Input } from '@base-ui/react'
-import { Plus, RefreshCw, Search } from 'lucide-react'
-import { useId, useState } from 'react'
-import type { ReactNode } from 'react'
-import { css } from 'styled-system/css'
-import * as v from 'valibot'
-
-import { StyledButton } from '../../../shared/components/styled-button'
-import { err, ok } from '../../../shared/domain/result'
-import type { Result } from '../../../shared/domain/result'
-import { fieldError, fieldInput } from '../../../styles/form'
-import type {
-  CreateTag,
-  CreateTagError,
-  CreateTagResult,
-  SelectableTag
-} from '../../tags/application/create-tag'
-import { tagNameSchema } from '../../tags/domain/tag-values'
-import type { TagId } from '../../tags/domain/tag-values'
-
-export type { CreateTag, CreateTagError, CreateTagResult, SelectableTag }
-
-type SelectionProps = {
-  readonly selectedTagIds: readonly TagId[]
-  readonly onSelectedTagIdsChange: (tagIds: readonly TagId[]) => void
-}
-
-const tagField = css({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '3',
-  margin: '0',
-  padding: '0',
-  borderWidth: 'none',
-  minInlineSize: '0'
-})
-
-const tagFieldLegend = css({
-  fontWeight: 'semibold',
-  marginBlockEnd: '1'
-})
-
-const tagOptionList = css({
-  display: 'flex',
-  flexDirection: 'column',
-  gap: '2',
-  maxBlockSize: '16rem',
-  overflowY: 'auto'
-})
-
-const tagOption = css({
-  display: 'flex',
-  alignItems: 'center',
-  gap: '2',
-  minBlockSize: 'touch'
-})
-
-const tagQueryRow = css({
-  display: 'flex',
-  flexWrap: 'wrap',
-  gap: '2',
-  alignItems: 'center',
-  '& > :first-child': {
-    flex: '1',
-    minInlineSize: '12rem'
-  }
-})
-
-const tagStatus = css({
-  margin: '0',
-  color: 'fg.muted'
-})
-
-const tagActions = css({
-  display: 'flex',
-  flexWrap: 'wrap',
-  gap: '2'
-})
-
-function Frame({ children }: { readonly children: ReactNode }) {
-  return (
-    <fieldset
-      className={tagField}
-      aria-label='タグ'>
-      <legend className={tagFieldLegend}>タグ</legend>
-      {children}
-    </fieldset>
-  )
-}
-
-function createErrorMessage(error: CreateTagError): string {
-  switch (error.code) {
-    case 'invalid-tag-name': {
-      return 'タグ名を入力してください（32文字以内）'
-    }
-    case 'duplicate-tag-name': {
-      return 'そのタグ名は既に存在します'
-    }
-    case 'unexpected-error': {
-      return 'タグの作成に失敗しました'
-    }
-    default: {
-      return 'タグの作成に失敗しました'
-    }
-  }
-}
-
-function parseCreateName(raw: string): Result<string, CreateTagError> {
-  const parsed = v.safeParse(tagNameSchema, raw)
-  if (!parsed.success) {
-    return err({ code: 'invalid-tag-name', field: 'name' })
-  }
-  return ok(parsed.output)
-}
-
-async function runCreateTag(
-  query: string,
-  onCreateTag: CreateTag
-): Promise<Result<SelectableTag, CreateTagError>> {
-  const nameResult = parseCreateName(query)
-  if (!nameResult.ok) {
-    return nameResult
-  }
-  return onCreateTag(nameResult.value)
-}
-
-function useCreateTagAction(onCreateTag: CreateTag, onCreated?: (tag: SelectableTag) => void) {
-  const [createError, setCreateError] = useState<string | null>(null)
-  const [creating, setCreating] = useState(false)
-
-  async function createFromQuery(query: string) {
-    setCreateError(null)
-    setCreating(true)
-    try {
-      const result = await runCreateTag(query, onCreateTag)
-      if (!result.ok) {
-        setCreateError(createErrorMessage(result.error))
-        return
-      }
-      onCreated?.(result.value)
-    } finally {
-      setCreating(false)
-    }
-  }
-
-  return {
-    createError,
-    creating,
-    clearCreateError: () => setCreateError(null),
-    createFromQuery
-  }
-}
-
-function QueryAndCreate({
-  query,
-  onQueryChange,
-  onCreateTag,
-  onCreated,
-  onBeforeCreate,
-  inputId
-}: {
-  readonly query: string
-  readonly onQueryChange: (value: string) => void
-  readonly onCreateTag: CreateTag
-  readonly onCreated?: (tag: SelectableTag) => void
-  /** タグ作成の直前に呼ばれる。更新 server error を clear する用途を想定。 */
-  readonly onBeforeCreate?: (() => void) | undefined
-  readonly inputId: string
-}) {
-  const { createError, creating, clearCreateError, createFromQuery } = useCreateTagAction(
-    onCreateTag,
-    onCreated
-  )
-
-  return (
-    <div>
-      <label htmlFor={inputId}>
-        <Search
-          size={16}
-          aria-hidden
-        />{' '}
-        タグを絞り込む / 新規作成
-      </label>
-      <div className={tagQueryRow}>
-        <Input
-          id={inputId}
-          className={fieldInput}
-          value={query}
-          type='text'
-          placeholder='タグを絞り込む / 新規作成'
-          disabled={creating}
-          onValueChange={(value) => {
-            onQueryChange(value)
-            clearCreateError()
-          }}
-        />
-        <StyledButton
-          type='button'
-          onClick={() => {
-            onBeforeCreate?.()
-            createFromQuery(query).catch(() => {
-              /* CreateError で表示 */
-            })
-          }}
-          disabled={creating}>
-          <Plus
-            size={16}
-            aria-hidden
-          />{' '}
-          {creating ? '作成中…' : 'この名前で作成'}
-        </StyledButton>
-      </div>
-      {createError !== null ? (
-        <p
-          className={fieldError}
-          role='alert'>
-          {createError}
-        </p>
-      ) : null}
-    </div>
-  )
-}
-
-// Why?
-// 更新 server error のタグ分は BookmarkForm の summary へ混ぜず、タグ領域だけで表示する。
-// タグ選択や新規作成といったタグ領域内の操作でだけ clear する意味を、
-// 「BookmarkTagField が受け取って表示する」「タグ操作時に onClearServerError を呼ぶ」の
-// 二点でコードに表現する。BookmarkForm を経由させると URL/title 側の入力変更でも
-// Clear が発火してしまい、責務境界がぼやける。
-function ServerErrorNotice({ message }: { readonly message: string }) {
-  return (
-    <p
-      className={fieldError}
-      role='alert'>
-      {message}
-    </p>
-  )
-}
-
-function Loading() {
-  return (
-    <Frame>
-      <p
-        className={tagStatus}
-        aria-busy='true'
-        aria-live='polite'>
-        タグを読み込み中…
-      </p>
-    </Frame>
-  )
-}
-
-function ErrorState({
-  message,
-  onRetry
-}: {
-  readonly message: string
-  readonly onRetry: () => void
-}) {
-  return (
-    <Frame>
-      <p
-        className={fieldError}
-        role='alert'>
-        {message}
-      </p>
-      <div className={tagActions}>
-        <StyledButton
-          type='button'
-          onClick={onRetry}>
-          <RefreshCw
-            size={16}
-            aria-hidden
-          />{' '}
-          再試行
-        </StyledButton>
-      </div>
-    </Frame>
-  )
-}
-
-function Blank({
-  onCreateTag,
-  onCreated,
-  serverError,
-  onClearServerError
-}: {
-  readonly onCreateTag: CreateTag
-  readonly onCreated?: (tag: SelectableTag) => void
-  /** BookmarkEditor が保持する更新 server error のうち tags 側だけを受け取る */
-  readonly serverError?: string | null
-  readonly onClearServerError?: () => void
-}) {
-  const inputId = useId()
-  const [query, setQuery] = useState('')
-
-  return (
-    <Frame>
-      <p className={tagStatus}>タグがまだありません。名前を入れて作成できます。</p>
-      {serverError != null && serverError !== '' ? (
-        <ServerErrorNotice message={serverError} />
-      ) : null}
-      <QueryAndCreate
-        inputId={inputId}
-        query={query}
-        onQueryChange={setQuery}
-        onCreateTag={onCreateTag}
-        onBeforeCreate={onClearServerError}
-        onCreated={(tag) => {
-          setQuery('')
-          onCreated?.(tag)
-        }}
-      />
-    </Frame>
-  )
-}
-
-function Ready({
-  tags,
-  selectedTagIds,
-  onSelectedTagIdsChange,
-  onCreateTag,
-  serverError,
-  onClearServerError
-}: SelectionProps & {
-  readonly tags: readonly SelectableTag[]
-  readonly onCreateTag: CreateTag
-  readonly serverError?: string | null
-  readonly onClearServerError?: () => void
-}) {
-  const inputId = useId()
-  const [query, setQuery] = useState('')
-  const normalizedQuery = query.trim().toLowerCase()
-  const filteredTags = tags.filter((tag) => tag.name.toLowerCase().includes(normalizedQuery))
-
-  function toggleTag(tagId: TagId) {
-    // Why?
-    // タグ選択の変更は「BookmarkEditor が保持している更新 server error (tags)」を
-    // 意味的に無効化するので、変更前に server error を clear する。
-    // BookmarkForm の URL/title 入力変更とは分離した clear 経路にしている。
-    onClearServerError?.()
-    if (selectedTagIds.includes(tagId)) {
-      onSelectedTagIdsChange(selectedTagIds.filter((id) => id !== tagId))
-    } else {
-      onSelectedTagIdsChange([...selectedTagIds, tagId])
-    }
-  }
-
-  return (
-    <Frame>
-      {serverError != null && serverError !== '' ? (
-        <ServerErrorNotice message={serverError} />
-      ) : null}
-
-      <div className={tagOptionList}>
-        {filteredTags.map((tag) => {
-          const optionId = `${inputId}-tag-${tag.id}`
-          return (
-            <label
-              key={tag.id}
-              htmlFor={optionId}
-              className={tagOption}>
-              <input
-                id={optionId}
-                type='checkbox'
-                checked={selectedTagIds.includes(tag.id)}
-                onChange={() => toggleTag(tag.id)}
-              />
-              {tag.name}
-            </label>
-          )
-        })}
-      </div>
-
-      <QueryAndCreate
-        inputId={inputId}
-        query={query}
-        onQueryChange={setQuery}
-        onCreateTag={onCreateTag}
-        onBeforeCreate={onClearServerError}
-        onCreated={(tag) => {
-          setQuery('')
-          if (!selectedTagIds.includes(tag.id)) {
-            onSelectedTagIdsChange([...selectedTagIds, tag.id])
-          }
-        }}
-      />
-    </Frame>
-  )
-}
-
-/**
- * タグ領域の状態ごとの明示的バリアント。
- * loading / blank / error / ready を boolean の組み合わせで切り替えない。
- */
-export const BookmarkTagField = {
-  Loading,
-  Error: ErrorState,
-  Blank,
-  Ready
-}
